requirements.txt
pyserial==3.5
pymongo==4.6.1
flask==3.0.2 